Rename browsers to browser
